Restructure code to allow cleaning up warnings
Make the top-level item in the crate to be a lib and then add all
applications under `src/bin` so that they are automatically defined
as binaries that depend on the top level lib.

This allows removing lots of warnings from build time.
