+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "validator-rust"
+name = "validator"
 version = "0.1.0"
 edition = "2021"
 default-run = "validator"
@@ -51,16 +51,12 @@ test-routes = []
 
 [[bin]]
 name = "validator"
-path = "src/main.rs"
+path = "src/bin/validator.rs"
 
 [[bin]]
 name = "wallet-tool"
-path = "src/wallet_tool.rs"
-test = false
-bench = false
+path = "src/bin/wallet_tool.rs"
 
 [[bin]]
 name = "message-builder"
-path = "src/message_builder.rs"
-test = false
-bench = false
+path = "src/bin/message_builder.rs"
