+++ src/server/mod.rs
@@ -1,5 +1,5 @@
-mod error;
-pub(crate) mod routes;
+pub mod error;
+pub mod routes;
 
 use std::net::SocketAddr;
 
@@ -72,34 +72,4 @@ where
     .bind(ctx.bind_address())?
     .run()
     .await
-}
-
-#[cfg(test)]
-pub mod test_utils {
-    use actix_web::{
-        web::{self, Data},
-        App,
-    };
-
-    use crate::{key_manager, state::ValidatorStateAccess};
-
-    use super::{
-        routes::{
-            self, get_tx::get_tx, index::index, post_tx::post_tx, sign::sign_route, test::set_state,
-        },
-        RuntimeContext,
-    };
-
-    fn create_test_app<Context, KeyManager, T>(runtime_context: Context) -> App<T>
-    where
-        Context: RuntimeContext
-            + routes::sign::Config<KeyManager>
-            + ValidatorStateAccess
-            + Clone
-            + Send
-            + 'static,
-        KeyManager: key_manager::KeyManager + Clone + Send + 'static,
-    {
-        todo!()
-    }
 }
